requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-auth==2.23.4
python-telegram-bot==20.7
pandas==2.1.4
openpyxl==3.1.2
requests==2.31.0
